+++ src/python_modules/app/framework.h
@@ -3,4 +3,5 @@
 #define WIN32_LEAN_AND_MEAN             // Windows ヘッダーからほとんど使用されていない部分を除外する
 // Windows ヘッダー ファイル
 #include <windows.h>
+#define SPDLOG_USE_STD_FORMAT
 #include <spdlog/spdlog.h>