+++ src/soyokaze/commands/regexp/RegExpCommand.cpp
@@ -54,11 +54,11 @@ bool RegExpCommand::PImpl::GetRegex(tregex& regexObject)
 			return true;
 		}
 		if (mMatchLevel == Pattern::FrontMatch) {
-			regexObject = tregex(fmt::format(_T("{}.*$"), (LPCTSTR)mParam.mPatternStr));
+			regexObject = tregex(std::format(_T("{}.*$"), (LPCTSTR)mParam.mPatternStr));
 			return true;
 		}
 		if (mMatchLevel == Pattern::PartialMatch) {
-			regexObject = tregex(fmt::format(_T("^.*{}.*$"), (LPCTSTR)mParam.mPatternStr));
+			regexObject = tregex(std::format(_T("^.*{}.*$"), (LPCTSTR)mParam.mPatternStr));
 			return true;
 		}
 		return false;